back-end/requirements.txt
Flask
Flask-JWT-Extended
Flask-CORS
Flask-PyMongo
Flask-Bcrypt
python-dotenv
